refactor(contracts): reorganize contract components and remove unused files
