--- apps/web/src/app/[lang]/(main)/(unirefund)/parties/[partyName]/[partyId]/contracts/_components/types.ts
@@ -1,1 +1,0 @@
-export type ContractPartyName = "merchants" | "refund-points";
