+++ apps/web/src/app/[lang]/(main)/(unirefund)/parties/refund-points/[partyId]/contracts/_components/table-data.tsx
@@ -53,7 +53,7 @@ const contractsTableColumns = ({
       name: {
         prefix: `/parties/refund-points/${partyId}/contracts`,
         targetAccessorKey: "id",
-        suffix: "/contract",
+        suffix: "contract",
       },
     },
     icons: {
